Move summary and review blocks for each question to separated files and folder
