+++ src/components/Application/PageElements/index.ts
@@ -1,0 +1,3 @@
+import PageElements from './PageElements'
+
+export { PageElements }
