+++ src/components/Application/index.ts
@@ -1,9 +1,9 @@
 import ApplicationHeader from './ApplicationHeader'
 import ApplicationSelectType from './ApplicationSelectType'
 import Navigation from './Navigation'
-import SectionWrapper from './SectionWrapper'
+import SectionWrapper from './Sections/SectionWrapper'
 import ProgressBar from './ProgressBar'
-import PageElements from './PageElements'
+import PageElements from './PageElements/PageElements'
 
 export {
   ApplicationHeader,
