+++ src/formElementPlugins/index.tsx
@@ -1,15 +1,15 @@
-// import ApplicationViewWrapper from './ApplicationViewWrapper'
+import ApplicationViewWrapper from './ApplicationViewWrapper'
 import TemplateViewWrapper from './TemplateViewWrapper'
-// import SummaryViewWrapper from './SummaryViewWrapper'
+import SummaryViewWrapper from './SummaryViewWrapper'
 import pluginProvider from './pluginProvider'
 import ErrorBoundary from './ErrorBoundary'
 import JsonInput from './JsonInput'
 
 export {
-  // ApplicationViewWrapper,
+  ApplicationViewWrapper,
   ErrorBoundary,
   JsonInput,
   TemplateViewWrapper,
-  // SummaryViewWrapper,
+  SummaryViewWrapper,
   pluginProvider,
 }
